Added service package and moved GrammarService into it
